+++ client/src/app/dependency/dependency.component.ts
@@ -1,10 +1,14 @@
 import { Component, Input, ChangeDetectionStrategy } from '@angular/core';
+import { TranslateModule } from '@ngx-translate/core';
+import { MatTooltip } from '@angular/material/tooltip';
 
 @Component({
-  selector: 'app-dependency',
-  templateUrl: './dependency.component.html',
-  styleUrls: ['./dependency.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+    selector: 'app-dependency',
+    templateUrl: './dependency.component.html',
+    styleUrls: ['./dependency.component.scss'],
+    changeDetection: ChangeDetectionStrategy.OnPush,
+    standalone: true,
+    imports: [MatTooltip, TranslateModule]
 })
 export class DependencyComponent {