+++ client/src/app/repository/repository.component.ts
@@ -3,11 +3,19 @@ import { Repository, GithubNotification, Update } from '../interface/repository.
 import { MatDialog } from '@angular/material/dialog';
 import { NotificationsDialogComponent } from './notifications-dialog/notifications-dialog.component';
 import { UpdatesDialogComponent } from './updates-dialog/updates-dialog.component';
+import { VersionPipe } from '../pipes/version.pipe';
+import { IconLibraryPipe } from '../pipes/icon-library.pipe';
+import { IconDesignPipe } from '../pipes/icon-design.pipe';
+import { TranslateModule } from '@ngx-translate/core';
+import { MatIcon } from '@angular/material/icon';
+import { MatTooltip } from '@angular/material/tooltip';
 
 @Component({
-  selector: 'app-repository',
-  templateUrl: './repository.component.html',
-  changeDetection: ChangeDetectionStrategy.OnPush
+    selector: 'app-repository',
+    templateUrl: './repository.component.html',
+    changeDetection: ChangeDetectionStrategy.OnPush,
+    standalone: true,
+    imports: [MatTooltip, MatIcon, TranslateModule, IconDesignPipe, IconLibraryPipe, VersionPipe]
 })
 export class RepositoryComponent {
     @Input() repository: Repository;